+++ packages/club_web/src/routes/invite/[token]/+page.svelte
@@ -140,6 +140,7 @@
 
 <style>
   .page {
+    width: 100%;
     min-height: 80vh;
     display: flex;
     align-items: center;